Let's validate Gitea instead of GitHub
This changes the validation to make sure that users can access port 3000
on their classroom server. We trust that Gitea itself will work like any
other web application if we've gotten this far, so we just open up the
port and serve an image from that port.

TRAINTECH-1435 #resolved #time 2h